Developed a system based on LAIS algorithm
In this project I developed a system which takes a sequence of any
length and outputs the longest almost increasing subsequence (LAIS).
Using the algorithm of LAIS, I wrote a code in Java which was a semester
long project of my graduate level course “Analysis of Algorithms”. I
also calculate a time complexity of this algorithm using counters.